+++ app/database_creation/database.py
@@ -1,14 +1,8 @@
-# app/database_creation/database.py
-
 from sqlalchemy import create_engine, event
-from sqlalchemy.orm import sessionmaker, declarative_base, Session
+from sqlalchemy.orm import sessionmaker, declarative_base
 
 DATABASE_URL = "sqlite:///./quickcart.db"
-
-engine = create_engine(
-    DATABASE_URL,
-    connect_args={"check_same_thread": False}
-)
+engine = create_engine(DATABASE_URL, connect_args={"check_same_thread": False})
 
 @event.listens_for(engine, "connect")
 def _enable_fk(dbapi_conn, _):
@@ -17,12 +11,11 @@ def _enable_fk(dbapi_conn, _):
 SessionLocal = sessionmaker(bind=engine, autoflush=False, autocommit=False)
 Base = declarative_base()
 
-# Create tables (ensure models are imported before this runs)
-Base.metadata.create_all(bind=engine)
 
-# Dependency for FastAPI
+# ...existing code...
+
 def get_db():
-    db: Session = SessionLocal()
+    db = SessionLocal()
     try:
         yield db
     finally:
