+++ app/models/models.py
@@ -1,0 +1,19 @@
+from sqlalchemy import Column, Integer, String, Float, ForeignKey
+from sqlalchemy.orm import relationship
+from app.database_creation.database import Base
+
+class Category(Base):
+    __tablename__ = "categories"
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String(100), unique=True, index=True)
+    products = relationship("Product", back_populates="category", cascade="all, delete", passive_deletes=True)
+
+class Product(Base):
+    __tablename__ = "products"
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String(100), index=True)
+    description = Column(String, nullable=True)
+    price = Column(Float)
+    quantity = Column(Integer)
+    category_id = Column(Integer, ForeignKey("categories.id", ondelete="CASCADE"), nullable=False)
+    category = relationship("Category", back_populates="products")
