+++ app/routers/category_router.py
@@ -1,21 +1,23 @@
-from fastapi import FastAPI, Depends, HTTPException, Query,APIRouter
-from app.schemas.category_schema import Category,CategoryBase,CategoryCreate,CategoryOut
-from typing import Optional,List
-from app.database_creation.database import Session,get_db
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.orm import Session
+from typing import List
+from app.schemas.category_schema import CategoryCreate, CategoryOut
+from app.models.models import Category
+from app.database_creation.database import get_db
 
-app = APIRouter()
+router = APIRouter()
 
-@app.get("/categories", response_model=List[CategoryOut])
+@router.get("/categories", response_model=List[CategoryOut])
 def list_categories(db: Session = Depends(get_db)):
     return db.query(Category).all()
 
-@app.post("/categories", response_model=CategoryOut)
+@router.post("/categories", response_model=CategoryOut)
 def create_category(c: CategoryCreate, db: Session = Depends(get_db)):
     obj = Category(**c.dict())
     db.add(obj); db.commit(); db.refresh(obj)
     return obj
 
-@app.put("/categories/{category_id}", response_model=CategoryOut)
+@router.put("/categories/{category_id}", response_model=CategoryOut)
 def update_category(category_id: int, c: CategoryCreate, db: Session = Depends(get_db)):
     obj = db.get(Category, category_id)
     if not obj:
@@ -24,7 +26,7 @@ def update_category(category_id: int, c: CategoryCreate, db: Session = Depends(g
     db.commit(); db.refresh(obj)
     return obj
 
-@app.delete("/categories/{category_id}", response_model=CategoryOut)
+@router.delete("/categories/{category_id}", response_model=CategoryOut)
 def delete_category(category_id: int, db: Session = Depends(get_db)):
     obj = db.get(Category, category_id)
     if not obj:
