+++ app/schemas/product_schema.py
@@ -1,21 +1,6 @@
-
-from sqlalchemy import create_engine, Column, Integer, String, Float, ForeignKey, event
-from sqlalchemy.orm import sessionmaker, declarative_base, relationship, Session
 from pydantic import BaseModel
 from typing import Optional
 from app.schemas.category_schema import CategoryOut
-from app.database_creation.database import Base
-
-class Product(Base):
-    __tablename__ = "products"
-    id = Column(Integer, primary_key=True, index=True)
-    name = Column(String(100), index=True)
-    description = Column(String, nullable=True)
-    price = Column(Float)
-    quantity = Column(Integer)
-    category_id = Column(Integer, ForeignKey("categories.id", ondelete="CASCADE"), nullable=False)
-    category = relationship("Category", back_populates="products")
-
 
 class ProductBase(BaseModel):
     name: str
@@ -24,7 +9,8 @@ class ProductBase(BaseModel):
     quantity: int
     category_id: int
 
-class ProductCreate(ProductBase): pass
+class ProductCreate(ProductBase):
+    pass
 
 class ProductOut(ProductBase):
     id: int
